Move feed cacheing to cache mod; put entry retreival into ctx mng
